web: added homepage and defaultengine variables
homepage is used when web is executed without arguments (except -p). by
default it's set to "about:home". before this change, google.com was
loaded when no arguments were passed and this slowed the process a
little because you had to wait for google.com to load and you didn't
have an empty address bar to start typing right away.

defaultengine is used when no search engine is specified.

improved also handling of filenames and urls that start with protocol:
open them directly only if no search engine is specified.